Refactor allocations domain

+++ packages/Server/Application/Campaigns/CampaignVotes.cs
@@ -1,3 +1,3 @@
 namespace Solidarity.Application.Campaigns;
 
-public record CampaignVotes(decimal EndorsedBalance, decimal Balance, double ApprovalThreshold);
+public record CampaignVotes(double EndorsedBalance, double Balance, double ApprovalThreshold);
